+++ TCC_3_M.csproj
@@ -134,6 +134,12 @@
     <Compile Include="Login.Designer.cs">
       <DependentUpon>Login.cs</DependentUpon>
     </Compile>
+    <Compile Include="MenuFornecedores.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="MenuFornecedores.Designer.cs">
+      <DependentUpon>MenuFornecedores.cs</DependentUpon>
+    </Compile>
     <Compile Include="MenuUsuários.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -214,11 +220,11 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Registro_Dispositivos.cs">
+    <Compile Include="RegistroDispositivos.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Registro_Dispositivos.Designer.cs">
-      <DependentUpon>Registro_Dispositivos.cs</DependentUpon>
+    <Compile Include="RegistroDispositivos.Designer.cs">
+      <DependentUpon>RegistroDispositivos.cs</DependentUpon>
     </Compile>
     <Compile Include="CadastroAdmin.cs">
       <SubType>Form</SubType>
@@ -252,6 +258,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="Menu.resx">
       <DependentUpon>Menu.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="MenuFornecedores.resx">
+      <DependentUpon>MenuFornecedores.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="MenuHardware.resx">
       <DependentUpon>MenuHardware.cs</DependentUpon>
@@ -296,8 +305,8 @@
     <EmbeddedResource Include="Vinculacao.resx">
       <DependentUpon>Vinculacao.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Registro_Dispositivos.resx">
-      <DependentUpon>Registro_Dispositivos.cs</DependentUpon>
+    <EmbeddedResource Include="RegistroDispositivos.resx">
+      <DependentUpon>RegistroDispositivos.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Splash.resx">
       <DependentUpon>Splash.cs</DependentUpon>
